+++ rs/rosetta-api/src/lib.rs
@@ -14,3 +14,4 @@ pub mod transaction_id;
 pub const API_VERSION: &str = "1.4.10";
 pub const NODE_VERSION: &str = env!("CARGO_PKG_VERSION");
 pub const DEFAULT_TOKEN_SYMBOL: &str = "ICP";
+pub const DEFAULT_BLOCKCHAIN: &str = "Internet Computer";