+++ rs/rosetta-api/CHANGELOG.md
@@ -3,7 +3,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
 ## Unreleased
+
+### Added
+- `blockchain` command line flag that overrides the blockchain name in the network identifier.
 
 ## [1.6.0] - 2022-05-30
 ### Fixed